+++ ft_memchr.c
@@ -1,4 +1,4 @@
-void	*ft_memchr(const void *str, int c, int n) // n type is size_t
+void	*ft_memchr(const void *str, int c, size_t n)
 {
 	return ;
 }
