+++ ft_memcmp.c
@@ -1,4 +1,4 @@
-int	ft_memcmp(const void *str1, const void *str2, int n) // n type size_t
+int	ft_memcmp(const void *str1, const void *str2, size_t n)
 {
 	return ;
 }
