+++ ft_memmove.c
@@ -1,4 +1,4 @@
-void	*memmove(void *str1, const void *str2, int n) // n type is size_t
+void	*memmove(void *str1, const void *str2, size_t n)
 {
-	return;
+	return ;
 }
