+++ ft_strncmp.c
@@ -1,4 +1,4 @@
-int	ft_strncmp(const char *str1, const char *str2, int n) // n type is size_t
+int	ft_strncmp(const char *str1, const char *str2, size_t n)
 {
-	return;
+	return ;
 }
