+++ ft_strnstr.c
@@ -1,4 +1,4 @@
-char	*ft_strnstr(const char *haystack, const char *needle, int n) // n size_t
+char	*ft_strnstr(const char *haystack, const char *needle, size_t n)
 {
 	return ;
 }
